requirements.txt
transformers==4.36.0
starlette==0.27.0
setuptools==65.5.1
openai-whisper
whisper
numpy
urllib3==1.26.18
pydantic==1.10.13
fastapi==0.100.0
uvicorn[standard]==0.20.0
llama-index==0.9.4
python-multipart==0.0.6
langchain==0.2.6
langchain-aws==0.1.6
langchain-community==0.2.6
langchain-openai==0.1.8
email-validator==1.3.1
PyPDF2==3.0.1
pypdf==3.17.1
pydub==0.25.1
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.0
google-cloud-speech==2.18.0
boto3==1.34.51
asyncpg==0.27.0
pytz==2022.7.1
ffmpeg
scikit-learn==1.2.1
marqo==2.1.0
redis>=5.0.1
docx2txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Pillow==10.2.0
psycopg2
google-generativeai
librosa
soundfile
SpeechRecognition
python-dotenv
langchain-google-genai